text_render: Support per-glyph overrides
Previously, an override font had to contain all the 256 glyphs in the
row that it was overriding. Experiments have shown that this would
cause huge file increase in CJK scripts (+200MiB or so).

Instead, allows the override font to only override some glyphs in a row.

For all glyphs (sprites) that have width 0, we now fall back to
the base font.